--- src/MUIPopover/MUIPopoverContent.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-class MUIPopoverContent extends React.Component {
-  render() {
-    return this.props.children;
-  }
-}
-
-export default MUIPopoverContent;
