--- src/MUIPopover/index.js
@@ -1,5 +1,0 @@
-import MUIPopover from "./MUIPopover";
-import MUIPopoverContent from "./MUIPopoverContent";
-import MUIPopoverTarget from "./MUIPopoverTarget";
-
-export { MUIPopover, MUIPopoverContent, MUIPopoverTarget };
